requirements.txt
alembic==1.7.5
cffi==1.15.0
click==8.0.3
colorama==0.4.4
cryptography==36.0.1
ecdsa==0.17.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Migrate==1.8.0
Flask-Script==2.0.6
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
greenlet==1.1.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
itsdangerous==2.0.1
Jinja2==3.0.1
Mako==1.1.6
MarkupSafe==2.0.1
psycopg2==2.9.3
pyasn1==0.4.8
pycparser==2.21
pycryptodome==3.6.6
python-dotenv==0.19.2
python-jose==3.3.0
rsa==4.8
six==1.16.0
SQLAlchemy==1.4.29
typing_extensions==4.0.1
Werkzeug==2.0.1
WTForms==3.0.1
zipp==3.7.0
